Renamed some terms and functions for greater consistency:
A "zone" is any data in an x,y,z grid of any size.
A "region" is a Minecraft region in any format.
An "mcmap" is a data structure containing a zone with blocks, a zone with
light levels, a zone with skylight levels, a 1-high zone containing a
height map, and a zone with opacity data.